Implement infinite scroll for top tracks
After investigating the API, I noticed it did not provide any information to
help on the pagination feature. The solution I came up with was to verify if the
track list returned from the API had any tracks in it, if it didn't then the
track list was over.

Still no scroll for the favorite tracks, the loading on the favorite pages is
preparing for future verification for the infinite scroll.