Implement logout feature
The DZ.logout function does not work as documented. First off, it doesn't accept
a callback argument, so there's no reliable way to know when the request to the
server has finished or even if it has finished successfully. The second issue is
that it doesn't clear the client cookie.

Therefore, this commit uses a new dependency called js-cookie to reset the
user's session.