Implement playlist feature
This commit creates a state for the playlist tracks that is a hashmap with
dynamic keys that will be constructed based on the playlist ID that has been
fetched.

I changed the payload on the playlist track list API to resolve with a map so I
could have access to the playlist ID, but I am now aware that using the action
meta is the most practical way to access it.

React Router has a function to deal with accessing the dynamic url that is
called useParams().